lms_project/requirements.txt
Django==2.2
djangorestframework==3.10.3
mysqlclient==2.0.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
